+++ api/interface.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/elodina/go-avro"
 	"github.com/gin-gonic/gin"
 )
 
@@ -36,9 +35,4 @@ type HyperBlockFacadeHandler interface {
 type HyperBlockProxy interface {
 	GetHyperBlockByNonce(c *gin.Context)
 	GetHyperBlockByHash(c *gin.Context)
-}
-
-// AvroEncoder should be able to encode any avro schema in a byte array
-type AvroEncoder interface {
-	Encode(record avro.AvroRecord) ([]byte, error)
 }
