+++ cmd/proxy/config/config.go
@@ -8,10 +8,10 @@ import (
 
 // Config holds the config for covalent proxy
 type Config struct {
-	Port              uint32 `toml:"Port"`
-	HyperBlockPath    string `toml:"HyperBlockPath"`
-	ElrondProxyUrl    string `toml:"ElrondProxyUrl"`
-	RequestTimeOutSec uint64 `toml:"RequestTimeOutSec"`
+	Port              uint32 `toml:"port"`
+	HyperBlockPath    string `toml:"hyperBlockPath"`
+	ElrondProxyUrl    string `toml:"elrondProxyUrl"`
+	RequestTimeOutSec uint64 `toml:"requestTimeOutSec"`
 }
 
 // LoadConfig will load the Config from the provided file
